# swagger/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Documentation",
    "description": "This is the API documentation for VMess project",
    "version": "1.0.0"
  },
  "host": "localhost:1002",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/updateProfilePicture": {
      "post": {
        "tags": [
          "User Profile"
        ],
        "summary": "Cập nhật ảnh hồ sơ",
        "description": "Endpoint để cập nhật ảnh hồ sơ người dùng",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/updateUserName": {
      "put": {
        "tags": [
          "User Profile"
        ],
        "summary": "Cập nhật tên người dùng",
        "description": "Endpoint để cập nhật tên người dùng",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/changePassword": {
      "post": {
        "tags": [
          "User Profile"
        ],
        "summary": "Đổi mật khẩu",
        "description": "Endpoint để đổi mật khẩu cho người dùng.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "oldPassword": {
                  "example": "any"
                },
                "newPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}